Add basic implementation for switch value accessor
The new toggle switch has a basic implementation of the controll value accessor. The switch can also handle ngModel inputs and registers changes.

The swtich takes two options 'on' and 'off'. The values set to the switch are constantly checked, so that they match either off or on. This allows the switch to only output and input the two allowed values. If a wrong value is beeing set to the switch, an error gets thrown.