Add base module for basic component inputs
The common input part of some components has been moved into a seperate module for inputs. The base components handle inputs like color, theme or size. There are some combinations of those. 

The base components can then be imported and extended. The inputs do not need to be declared anymore.