Add disabled state to toggle switch
The control value accessor needs more functionality than other components. There is a new method called setDisabled, which sets the disabled state of the component. The other possibility is to set the state via an input called disabled.

When disabled, the component adds a class called disabled to the toggle switch root div.