Rafactor control value accessor into base class
The control value accessor used by the radio and checkbox group is now in a base class, which handles the output and input value of the control. The base class then triggers an update signal, when the innerValue gets set. There are two separate events: outerChange and innerchange. The innerChange write method sets the innerChange value to true. This can then be used to avoid multiple loops.